@@ -1,96 +1,96 @@
 {
   "name": "typescript-react-native-starter",
-  "version": "0.0.4",
+  "version": "0.0.5",
   "private": false,
   "description": "A Typescript React Native boilerplate to start your project quickly",
   "author": "Hung Arthur Dao <arthur_dao@icloud.com>",
   "license": "MIT",
   "scripts": {
+    "tsc": "./node_modules/typescript/bin/tsc",
     "start": "node node_modules/react-native/local-cli/cli.js start",
     "start:ios": "open ./ios/*.xcworkspace && yarn start",
     "open:ios": "open ./ios/*.xcworkspace",
     "open:android": "open -a '/Applications/Android Studio.app'",
     "start:android": "open -a '/Applications/Android Studio.app' ./android && yarn start",
-    "lint": "eslint '*/**/*.{js,ts,tsx}'",
+    "lint": "yarn tsc && eslint . --ext .js,.ts,.tsx --cache",
     "test": "jest --verbose --coverage",
     "test:update": "jest --verbose --coverage --updateSnapshot",
     "test:watch": "jest --verbose --watch",
     "coverage": "jest --verbose --coverage && open ./coverage/lcov-report/index.html",
-    "codecov": "jest --verbose --coverage && ./node_modules/.bin/codecov --token=176f4867-f830-40ec-a536-90e7b611d713",
     "rename": "react-native-rename",
     "jetify": "npx jetify",
     "icon": "npx app-icon generate -i publishing/icon.png",
-    "setup": "./bin/install.sh $1"
+    "splash": "npx generate-bootsplash"
+  },
+  "dependencies": {
+    "@react-native-community/async-storage": "^1.12.0",
+    "@react-native-community/masked-view": "^0.1.10",
+    "@react-navigation/bottom-tabs": "^5.9.0",
+    "@react-navigation/native": "^5.7.4",
+    "@testing-library/react-native": "^7.0.2",
+    "@types/react-native-vector-icons": "^6.4.6",
+    "@types/react-redux": "^7.1.9",
+    "autobind-decorator": "^2.4.0",
+    "i18next": "^19.8.2",
+    "nock": "^13.0.4",
+    "react": "16.13.1",
+    "react-i18next": "^11.7.3",
+    "react-native": "0.63.2",
+    "react-native-bootsplash": "^2.2.7",
+    "react-native-easy-icon": "^1.0.6",
+    "react-native-gesture-handler": "^1.8.0",
+    "react-native-localization": "^2.1.6",
+    "react-native-normalize": "^1.0.1",
+    "react-native-reanimated": "^1.13.1",
+    "react-native-safe-area-context": "^3.1.8",
+    "react-native-screens": "^2.11.0",
+    "react-native-vector-icons": "^7.1.0",
+    "react-redux": "^7.2.1",
+    "redux-persist": "^6.0.0",
+    "redux-saga": "^1.1.3",
+    "redux-saga-test-plan": "^4.0.0-rc.3",
+    "typesafe-actions": "^5.1.0"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.11.6",
+    "@babel/runtime": "^7.11.2",
+    "@react-native-community/eslint-config": "^2.0.0",
+    "@types/jest": "^26.0.14",
+    "@types/react": "^16.9.49",
+    "@types/react-native": "^0.63.22",
+    "@types/react-test-renderer": "^16.9.3",
+    "@typescript-eslint/eslint-plugin": "^4.2.0",
+    "@typescript-eslint/parser": "^4.2.0",
+    "app-icon": "^0.13.2",
+    "babel-jest": "^26.3.0",
+    "babel-plugin-module-resolver": "^4.0.0",
+    "eslint": "^7.9.0",
+    "eslint-config-prettier": "^6.12.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jest": "^24.1.0",
+    "eslint-plugin-prettier": "^3.1.4",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-react": "^7.21.4",
+    "eslint-plugin-react-hooks": "^4.1.2",
+    "eslint-plugin-react-native": "^3.10.0",
+    "eslint-plugin-react-native-a11y": "^2.0.3",
+    "eslint-plugin-testing-library": "^3.9.2",
+    "husky": "^4.3.0",
+    "jest": "^26.4.2",
+    "jetifier": "^1.6.6",
+    "metro-react-native-babel-preset": "^0.63.0",
+    "prettier": "^2.1.2",
+    "react-dom": "^16.13.1",
+    "react-native-rename": "^2.4.1",
+    "react-native-typescript-transformer": "^1.2.13",
+    "react-test-renderer": "16.13.1",
+    "ts-jest": "^26.4.1",
+    "typescript": "^4.0.3"
   },
   "husky": {
     "hooks": {
       "pre-commit": "yarn lint",
-      "pre-push": "yarn codecov"
+      "pre-push": "yarn jest --coverage --changedSince=origin/master"
     }
-  },
-  "dependencies": {
-    "@react-native-community/async-storage": "^1.9.0",
-    "@react-native-community/masked-view": "^0.1.9",
-    "@react-navigation/bottom-tabs": "^5.2.6",
-    "@react-navigation/native": "^5.1.5",
-    "@types/react-redux": "^7.1.7",
-    "autobind-decorator": "^2.4.0",
-    "react": "16.11.0",
-    "react-native": "0.62.2",
-    "react-native-easy-icon": "^1.0.2",
-    "react-native-gesture-handler": "^1.6.1",
-    "react-native-localization": "^2.1.6",
-    "react-native-normalize": "^1.0.1",
-    "react-native-reanimated": "^1.8.0",
-    "react-native-safe-area-context": "^0.7.3",
-    "react-native-screens": "^2.5.0",
-    "react-native-vector-icons": "^6.6.0",
-    "react-redux": "^7.2.0",
-    "redux-persist": "^6.0.0",
-    "redux-saga": "^1.1.3",
-    "typesafe-actions": "^5.1.0"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.9.0",
-    "@babel/runtime": "^7.9.2",
-    "@react-native-community/eslint-config": "^1.0.0",
-    "@types/enzyme-adapter-react-16": "^1.0.6",
-    "@types/jest": "^25.2.1",
-    "@types/react": "^16.9.34",
-    "@types/react-native": "^0.62.2",
-    "@types/react-test-renderer": "^16.9.2",
-    "@typescript-eslint/eslint-plugin": "^2.28.0",
-    "@typescript-eslint/parser": "^2.28.0",
-    "app-icon": "^0.13.2",
-    "babel-jest": "^25.3.0",
-    "codecov": "^3.6.5",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.2",
-    "eslint": "^6.8.0",
-    "eslint-config-prettier": "^6.10.1",
-    "eslint-plugin-prettier": "^3.1.3",
-    "husky": "^4.2.5",
-    "jest": "^25.3.0",
-    "jetifier": "^1.6.5",
-    "metro-react-native-babel-preset": "^0.59.0",
-    "prettier": "^2.0.4",
-    "react-dom": "^16.13.1",
-    "react-native-rename": "^2.4.1",
-    "react-native-testing-library": "^1.13.0",
-    "react-native-typescript-transformer": "^1.2.13",
-    "react-test-renderer": "16.11.0",
-    "ts-jest": "^25.3.1",
-    "typescript": "^3.8.3"
-  },
-  "jest": {
-    "preset": "react-native",
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js",
-      "jsx",
-      "json",
-      "node"
-    ]
   }
 }
