@@ -1,10 +1,8 @@
 declare module 'redux-persist/lib/*';
 
-export interface BaseAction {
-  type: string;
-  payload: any;
-}
-
 export interface User {
-  id?: string;
+  id: string;
+  email: string;
+  phone: string;
+  name: string;
 }
